sdk: tests: refactor: move fee models unit tests
